Test account API get method and authentication middleware
Basic implementation of the account API endpoint, focusing on the GET method.
The method shall return the user object belonging to the currently signed in user,
by fetching the user ID with JSON Web Tokens.

Before this commit, JWT was not implemented. Thus, a basic implementation is done is
this commit as well.

See DIH-100